\name{tf.list}
\alias{tf.list}
\title{The candidate pool of potential regulators}
\description{
tf.list was constructed by intersecting the periodic genes and the 
transcription factors in YEASTRACT \url{http://www.yeastract.com/}.
}
\usage{data(tf.list)}
\format{A list of potential regulators}
\source{
\url{http://www.biology.duke.edu/haaselab/publicData/index.html}
\cr \cr
\url{http://www.yeastract.com/}
}
\references{
\cite{Orlando, D.A., Lin, C.Y., Bernard, A., Wang, J.Y., Socolar, J.E., Iversen, E.S., Hartemink, A.J. and Haase, S.B. (2008) Global control of cell-cycle transcription by coupled CDK and network oscillators, Nature, 453, 944-947.}
\cr \cr
\cite{Teixeira, M.C., Monteiro, P., Jain, P., Tenreiro, S., Fernandes, A.R., Mira, N.P., Alenquer, M., Freitas, A.T., Oliveira, A.L. and Sa-Correia, I. (2006) The YEASTRACT database: a tool for the analysis of transcription regulatory associations in Saccharomyces cerevisiae, Nucleic acids research, 34, D446-451.}
}


