\name{wt.expr.data}
\alias{wt.expr.data}
\title{Time course cell cycle dataset of wild-type cells}
\description{
wt.expr.data is from a yeast cell cycle datasets which is publicly available at GEO 
\url{http://www.ncbi.nlm.nih.gov/geo} under accession number GSE8799. 
The time course gene expression profiles included 15 time points 
at 16 min resolution in wild type Saccharomyces cerevisiae.
}
\usage{data(wt.expr.data)}
\format{1267 genes expression profiles at 15 time points}
\source{\url{http://www.biology.duke.edu/haaselab/publicData/index.html}}
\references{
\cite{Orlando, D.A., Lin, C.Y., Bernard, A., Wang, J.Y., Socolar, J.E., Iversen, E.S., Hartemink, A.J. and Haase, S.B. (2008) Global control of cell-cycle transcription by coupled CDK and network oscillators, Nature, 453, 944-947.}
}


