\name{yeastract}
\alias{yeastract}
\title{yeastract}
\description{
YEASTRACT are the potential gene regulation relationship.
}
\usage{data(yeastract)}
\format{2735 potential gene regulation pairs, the first column is regulator,the second column is target gene}
\source{\url{http://www.yeastract.com/}}
\references{
\cite{Teixeira, M.C., Monteiro, P., Jain, P., Tenreiro, S., Fernandes, A.R., Mira, N.P., Alenquer, M., Freitas, A.T., Oliveira, A.L. and Sa-Correia, I. (2006) The YEASTRACT database: a tool for the analysis of transcription regulatory associations in Saccharomyces cerevisiae, Nucleic acids research, 34, D446-451.}

\cite{Monteiro, P.T., Mendes, N.D., Teixeira, M.C., d'Orey, S., Tenreiro, S., Mira, N.P., Pais, H., Francisco, A.P., Carvalho, A.M., Lourenco, A.B., Sa-Correia, I., Oliveira, A.L. and Freitas, A.T. (2008) YEASTRACT-DISCOVERER: new tools to improve the analysis of transcriptional regulatory associations in Saccharomyces cerevisiae, Nucleic acids research, 36, D132-136.}
}


