\name{timedelay.lm.batch}
\alias{timedelay.lm.batch}
\title{time delay linear model batch}
\description{
Batch usage of time delay linear model.
}
\usage{
timedelay.lm.batch(expr,regulator2target, \dots)
}
\arguments{
  \item{expr}{an expression matrix with genes in the rows, time points in the columns}
  \item{regulator2target}{potential gene regulation relationship}
  \item{ \dots}{Additional arguments, passed to \link{timedelay.lm}}
}

